+++ CHANGELOG.md
@@ -7,23 +7,40 @@ All notable changes to this project will be documented in this file.
 
 Releases will be numbered with the following format:
 
-`<major>.<minor>.<patch>`
+**`<major>.<minor>.<patch>`**
 
 And constructed with the following guidelines:
 
-1. MAJOR version when you make incompatible API changes **bumps the major** resets minor and patch
-2. MINOR version when you add functionality in a backwards-compatible manner **bumps the minor** resets patch
-3. PATCH version when you make backwards-compatible bug fixes and misc changes **bumps only the patch**
+1. **MAJOR** version when you make incompatible API changes **bumps the major** resets minor and patch
+2. **MINOR** version when you add functionality in a backwards-compatible manner **bumps the minor** resets patch
+3. **PATCH** version when you make backwards-compatible bug fixes and misc changes **bumps only the patch**
+
 =========================================================================================================
 
 ## Outstanding Items
 
 #### Items to Add
-* Search box
 * Application Cache Manifest - http://www.html5rocks.com/en/tutorials/appcache/beginner/
 * geo tracking
 
 ===========================================================================================================
+
+## Version 3.2.0 | 01/12/2017
+
+* added Phoenix Sonoran Bikeway to map
+* updates to Bikeways data
+* updates to GRIDshare data
+* update to [ArcGIS API for JavaScript 3.19compact] (https://developers.arcgis.com/javascript/index.html)
+
+## Version 3.1.3 | 06/23/2016
+
+* added a help page
+* added a search function
+* changed menu icon from Gear to Info
+* added USBR90 to map
+* Updated Light Rail Line
+* Updated Transit Locations
+* updated Bike Shop info
 
 ## Version 3.1.2 | 06/08/2016